chore: use long return type

+++ backend/src/main/java/format/backend/submission/dto/SubmissionAnswersStatisticsResponseDto.java
@@ -1,3 +1,3 @@
 package format.backend.submission.dto;
 
-public record SubmissionAnswersStatisticsResponseDto(String answerId, Integer totalCount) {}
+public record SubmissionAnswersStatisticsResponseDto(String answerId, long totalCount) {}
